requirements.txt
# ============================================
# CORE FLASK & WEB FRAMEWORK
# ============================================
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0

# ============================================
# DATABASE
# ============================================
mysql-connector-python==8.1.0

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
bcrypt==4.1.2

# ============================================
# SESSIONS & STORAGE
# ============================================
Flask-Session==0.5.0

# ============================================
# REAL-TIME COMMUNICATION (CHAT)
# ============================================
# Hapus versi spesifik untuk menghindari conflict
Flask-SocketIO>=5.3.0
python-socketio>=5.0.0
eventlet>=0.35.0

# ============================================
# PAYMENT GATEWAY & HTTP REQUESTS
# ============================================
requests==2.31.0

# ============================================
# FILE UPLOAD & IMAGE PROCESSING
# ============================================
Pillow==10.1.0

# ============================================
# DATE & TIME
# ============================================
python-dateutil==2.8.2

# ============================================
# ENVIRONMENT VARIABLES
# ============================================
python-dotenv==1.0.0

# ============================================
# PAYMENT GATEWAY & HTTP REQUESTS
# ============================================
requests==2.31.0
midtransclient==1.4.1